backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
libretranslatepy==2.1.1
lxml==4.9.1
pytz==2022.4
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.3
translate==3.6.1
tzdata==2022.5
urllib3==1.26.12
